Fix error during pipe entry
As Mario is sliding down the pipe, he leaves the Area2D that detects if he's on the pipe. This triggers `_on_mario_off(...)`, which nulls the reference to Mario...while the pipe is still sliding him.

Easiest fix is to avoid nulling the Mario reference when the slide is in progress, but it also works to extend the Area2D downwards so that he doesn't even leave it before being teleported. I've applied both fixes.