Create exit_pipe.gd
Instead of sending you somewhere in the current room, sends you to a different scene.

This is definitely not how the Door class does it, but this feels more elegant. It seems disingenuous to give all pipes a reference to Singleton/WindowWarp just so a select few of them can use it.

(Admittedly, I don't know if doing that would cause performance issues. Godot is new territory for me.)