Use global position to place Mario
Vexxter discovered that when signs are children of other nodes, the parent nodes' positions are not taken into account when shifting Mario over to read the sign. This means that if you load in a scene then move its root node, reading a sign in that scene will teleport Mario somewhere crazy.

Using the interactable's global position takes parent nodes into account, thereby solving the problem.