Use read_pos_x instead of manual lerp
For consistency with interactable_dialog. Includes a note on the difference between this method and the old method, for future reference.

Changing the order of operations in Door._update_animation(...) fixes a bug in this new centering implementation where Mario always slides to the center of double doors, rather than to the appropriate half. This happened because the choose-a-side code was being run AFTER the send-Mario-to-one-side code, and because the new impl. only runs once at the animation's beginning, it wasn't getting fixed on future frames.