Move door anim properties into skin resources
The idea is that the animation properties will be tuned to fit a particular sprite flipbook, therefore the properties should be tied to that flipbook.

I'm impressed, because this would've been way harder in Unity.
I am a bit ticked that it doesn't let you export custom resource types though...that's a little annoying.