Fix bug where door loses sprite on scene load
It was very annoying. When you'd open the scene in the editor, the door's Sprite field would get unset and the game would refuse to run until you manually reverted it to the default (set) value.

Problem came from running the script as tool, it turns out. At present, Interactable (the base-st version) has an `onready` variable that assigns the sprite field in some weird hootenany that I don't understand.

IIRC Creyon thinks that part of Interactable could be moved to InteractableDialog, but because this is a door PR, I'd rather not touch that part. As such, here's a fix that just gets rid of the door's tool status and moves the graphic's setget behavior into _ready().