Tidier set_disabled override
Occurs to me, this definitely was still being used. Same name as the root and everything.

Anyway, ride_area was just the area that checks for player overlap...aka the object's root node. And it looks like Interactable takes care of setting monitoring, meaning the only important(?) new part is setting the collider's collision layer bit.