Make fuse quieter after bomb's been hit
Just a little QoL feature. The fuse has a very wide hearing distance by default, which is great when you need be aware of the Bob-Omb's location--but once the bomb's been hit away, you don't care about it anymore, so having its fuse still audible from a long distance just feels strange.

The subtleties of this feature are documented in the code.