Fix silent pound on edges and steep slopes
Because the step raycast only checks 20 units down from dead center, it was liable to draw a blank if the player was a good ways off a platform's edge, or on a steep slope.

Fixed by checking raycasts on either side of the player's hitbox as well as the one in dead center. This way, the only way for a cast to miss is if the player is literally not making contact with ground (or if the ground is a sharp peak hitting 1/4 or 3/4 of the way across...ah, the foibles of raycasting :smirk:).