Start burnaway early to avoid lag
If the burnaway lags on the final frame, the transition looks pretty awful--freezing on a frame with a little painting left, then smashing to a frame where the scene has already started reappearing, meaning there's no frames of solid white in between.

Easiest way to avoid this is to have the burnaway end before scene loading is even thinking of beginning.