Only let one painting ripple at a time
Having the material constantly attached to all paintings, as I feared, leads to every painting rippling in sync. Easiest fix is to just keep most paintings' materials unset by default, and only load it onto the ones which are about to come active.

If multiplayer is needed, each painting could instantiate a copy of the material before it uses it. This works fine for now, though.