Preprocess to find all interactables early
The "more elegant" approach I alluded to is this: on ready, find all nodes in the options menu that are any kind of button (meaning anything you can click on) and add them to an array. Once that's done, enabling and disabling all interactables is as easy as iterating that array.

Should be more efficient, too. Enabling and disabling the options menu isn't a terribly demanding process (won't be run every frame except maybe in a TAS, part of the pause menu rather than the simulation loop), but even so.

#2 is now properly fixed.