Resequence death event
Death cover has no reference to the player. We could just have it reference the player, but the point of refactoring like this is to make multiplayer work--and I don't think referencing one specific player counts as multiplayer-compatible!

A more reasonable design (in my opinion) is to flip the flow of action. Instead of death cover calling Singleton to check if the player is dead, Singleton calls the death cover when the player is dead. This also means we can do more complicated logic in Singleton, e.g. only bringing the death cover in when all players are dead simultaneously.

Granted, this isn't objectively superior than the all-players-dead-at-once logic being in the death cover. I think it's just...different. It's all to do with the single responsibility principle, I guess. The objective best solution might be to move the death-counting logic into its own class, actually.