Bring back death system redesign
Unfortunately, death cover is created on the title screen, before the player exists. Need a system that doesn't expect the player to exist always...and in a way, I suppose this is it.

The main problem, I suppose, is that death cover is actually two systems: the death system, and the fade-out system. What I really need is to split those up!