Ctor so locking objects from code is easy
The disadvantage of the "enclosing node" lock model is difficulty of setup. BUT--said setup can be automated, at least to a certain extent!

Positioning the lock graphic and hitbox will still be an annoyance, but hopefully it'll be easier if the node getting/setting isn't also being done by hand.