Use common enemy sound hooks for Koopas
Doesn't really solve any problems, but makes the codebase more consistent, I think.

The script is set up to use two of the three built-in sound hooks, but for now these are both hooked to the same sound source. This is just in case we want to have different stomp and struck sounds in the future...though I can't honestly see us needing that.