Upgrade double-connect to soft error
Error was found in koopas, which it turns out call ResidualSFX.new_from_existing twice per sfx (once in their own code, once in their EnemyWalk base class). Calling twice is a definite less-desirable thing, so it should be reported. Not a super big deal, though, since it's not the end of the world if it does happen (hence, soft error).