Add animation syncing
For some player anims (e.g. crouching), the most practical way to define FLUDD's poses is to just create a matching anim and have it play in sync with the player. This commit adds support for that (in place of Y flipping, which was never going to see use anyway).