Comment and cleanup pencil.gd
This is where the bulk of the terrain craziness happens, so it's important to understand it. Trying to keep actual logic changes to a minimum, but I see nothing wrong with renaming unclear variables (including in other files :P). And of course, comments and type hints are a no-brainer.

Also: left comments for myself on what larger changes I'd like to make in future commits.