Deduplicate most duplicated code
Much of the polygon generation code is exactly the same between top, sides, and bottom--sides and bottom were a few values off from exactly identical, in fact! Top is a good bit different, but it does still share a decent chunk of vertex-prep and display code.

So I moved the shared bits into their own functions because of course :P

There is a bit of duplication left--I'm not sure how to dedup top-area intersection code, since it's so simple.