Rename supporting terrain scripts
...to make it more obvious what they do.

"Pencil" implies a level designer tool, which it absolutely isn't. "Terrain border" is a little more self-explanatory.

And as for "endcap," it's a lot less vague than "edge." Just in the terrain system alone, "edge" could already mean three things (edge of the source polygon; edge of a generated border polygon; displayed edges of the top grass), so it wasn't clear at a glance what the script did. "Endcap," on the other hand, is unambiguous.