Eval all type IDs before drawing any edges
Before, edge-drawing code would evaluate and save type IDs as it went, so that by the end of the top- and bottom-drawing code, most edges had been evaluated. The problem was, this made draw code order-dependent--top and bottom HAD to come before sides, so they had to use Z-order in order to appear in front.

By moving the dictionary eval into a pre-process, the drawing steps become order-independent, meaning front and back can be created using pure draw order.