Untangle interactable's sprite-finding line
It was weird in 3.5, but in 4.1 it was crashing!

Rather than try to make it work the same, I just expanded it out into a _ready() function and a second variable, then updated all Interactables to match. Been meaning to do that for a long time, actually--it made tool scripts unusable :)