Fix audio-related Goomba bug
A while back, saw a crash from a goomba's audio source. I've noticed a large chunk of data being updated out of the prefab; I guess that was actually packed Vorbis data, which may not be packable anymore, because the sound source which caused the crash had a completely broken reference.