Fix pausing and unpausing
Both the Pause Menu and the HUD have to operate in all states (paused or unpaused) for the menu to work. HUD has to work when paused because it handles unpausing; menu has to work when unpaused because it handles its own close animation.

The menu itself is borked pretty bad, but at least the game doesn't lock on attempting to pause anymore.