Rewrite and unfold distance field cutoff algorithm
The math looks a bit different, but it's mathematically equivalent, and gives a bit more intuitive understanding of what I added last time.

(An explanation for that, incidentally: I wanted the droplets to be expandable, but wanted the older droplets to visibly shrink just as much. Multiplying expand/reduce by the vertex alpha built into COLOR.a did just that.)