Reorder all main LD scripts to standard
Includes only the scripts found in the main editor scene's node hierarchy.

Scripts were updated so that variables are in the correct order (export, private, onready); so that functions are in the correct order (internal fns like _ready, _process, etc. before any others); and so that functions are spaced to standard (2 lines between, no fewer).