Have item pane populate item grid
I've had to modify the item grid before, and I'm tired of having to tease out *which* parent class populates it. Just move it into a node directly related to said grid.

Script is now self-starting, instead of having main start it. I'm sure this can be done better, but this appears to work for now.

This puts this branch squarely in the realm of refactoring.