Rewrite glow shader's bounds scaling
Old function scaled the mesh uniformly on each axis rather than adding an exact thickness to each side. The UV function didn't know this, causing terrible UV squishing as the outline grew thicker.

The new algorithm doesn't exhibit this problem, and is also branchless.

Also, kind of silly to recalculate the UV for every fragment. We're not using any per-fragment data in their calculations; let's just do that in the vertex shader and save the extra work.

The new algorithm is deeply flawed, unfortunately--Shine Sprites break it in no fewer than two ways.