Comment, type-hint, and cleanup property_menu.gd
Now here's a nice, simple set of changes.

The name "key" bugs me--it doesn't give any clues about what data is actually contained in the variable, which makes code difficult to understand. Thus the change to "propname" (given that fields' labels are being assigned its value, I'm inferring that that's what's in it--the property's name).