Make tickbox feel responsive
The existing system had all frames of the tickbox animation in a single animation, arranged so that starting the animation makes the checkmark appear. The problem is, it had to go all the way through the unchecked first frame--so it had a perceptible delay between click and visual response.

The simplest fix would be to jump to frame 1 when the animation starts. But if we're putting different animation code in the "if pressed" and "else" blocks, might as well go all in and use the traditional two-different-animations system.