Finally--working UV correction
I neglected to commit the non-working versions, but I struggled tooth and nail to get this cel index algorithm working. As is often the case, in the end it turned out that I'd made a simple mistake (adding outside the parentheses instead of inside), and once that was fixed the entire thing Just Worked.

Next step: I think I know how to make it work without having to pass cel index, even though it'll involve doing the UV correction in the fragment shader (so, per-pixel - slightly less than ideal).
UPDATE: I can't get textureSize to work--so for now, this is off the table. Sticking with the pass-cel-index-by-script method.