Fix buffer reading too few bytes
Aha! A crash which is explicitly from changed behavior in 4.1. The `slice` function is start-inclusive, end-exclusive--but the `subarray` function it replaced is *both-inclusive!* As such, the slice endpoint was being adjusted to account for a quirk that no longer existed--and every buffer read was coming out one byte shorter than intended. Thus, it was reading garbage data--and crashing.

Loading a blank file now works properly (or at least, doesn't crash). Loading a file with a single coin comes out blank. Loading a file with one of everything still crashes after reading a garbage ID.