Move mission-list fns near level fns that use'm
General programming convention is (apparently) to keep functions only used in one place under the function which calls them. I'm not breaking the interleaving to do that to the letter (again--with save and load interleaved, you can quickly verify that save and load handle the data identically), but I can still do it in spirit.