Fix music fading out
Weird! It seems like it was starting the fade-out tween at basically silence. Bet that was because the fade-in was starting, applying its effect to volume, and essentially passing that changed volume to fade-out for it to use as its starting value. (The fact that it's storing a starting value *at all* is a prime example of a 4.1 change breaking things.)

Setting fade-out's starting value explicitly fixes the bug.