Delete bg_plane.gd
No longer needed, since the combination of ParallaxLayer and ScrollingParallaxLayer can do everything we know we need. Infinite-on-one-side is not yet covered, but we don't even know that it's useful yet. And when it is needed, it's probably best to implement it as a separate feature (or just fake it with a really really long sprite).