Actually write flipped points to polygons
I forgot that iterating an array of vectors yields copies of vectors, not references. So every element was getting copied, the copy was being written to, and the loop was moving on with the original unchanged.

Mirrored rocks now have correct collision. Good that setter fns run when the game starts, even if they don't in the editor; the LD is still broken in this branch, so I had to manually add the rocks to the test level :stuck_out_tongue_closed_eyes: