Disable solids via process_mode, not layer mask
Disabling by layer mask relies on setting the correct layer mask, which introduces an extra point of failure (the object's layer mask can mismatch what's coded). Godot's built-in disable-physics-object system does not have this limitation, but it does have multiple available disablement behaviors. In this case, though, the most common option is the default, and unlike layer masks, only needs to be changed if the other behaviors are needed.

Even so, document the dependence on disable_mode in the class's docstring.