Fix LD crashing in BG code
Turns out that when I refactored backgrounds in #184, I removed a node which looked unnecessary, but which the LD was secretly using to darken the BG. As a result, since merging from master, the LD would crash immediately on game start.

To fix the crash, let's use a plain translucent ColorRect in place of a code-based darkening system which relies on the BG's node structure. It looks identical, and this way, future BG designers don't have to worry about One Arcane Node That Can Break The Game If You Don't Include It.