Rename things in terrain border generators
And since I'm renaming as well, I might as well replace the root node's texture variables with direct refs to the textures in the skin. (The reason I kept the variables was so I wouldn't have to replace all call sites in the border scripts, but. I'm retouching the border scripts anyway ( ͡° ͜ʖ ͡°))