Replace temp resource with imported spritesheet
Gone are these extra loose files. Now, it's just the one, as it was always meant to be.

It annoys me that terrain_polygon has gone back to storing a copy of the body texture, but hey, there's not much I can do about that. Apart from one thing, I guess.