Deduplicate tex-slice-creation code
This duplication is intolerable. Even with VSCode's multi-cursor feature, it was becoming impractical to add things to, and was already straining the limits of readability. I tried to add something new, and just went "nope."

Doing it this way also sets up for a potential future in which terrain skins can have arbitrarily many border textures. Not in this PR though.