Add option to store terrain textures separate
Technically, this is the last key ingredient needed for the importer to be capable of everything either system could do. I'd like to add a system to detect when a texture has changed separate from the importer, to avoid overwriting people's work, but the system is technically usable at this point.