Cache joypad button names statically
This drops processing time from ~4.5 ms down to ~0.25 ms. This puts it level with Entity._physics_process in T1R1. Frankly a bit shocking that an options menu should even be that intensive, though.

The script already has machinery to mostly operate without _process(). Looking into how to take that to 100% so we can skip that function entirely.