Only update level_info font when locale changes
Perhaps it's not taking 11 milliseconds, but this script's process function was up there with water and entities last I checked--which, again, is pretty outrageous for a static pause menu element! Now it only updates when the language changes, not once every frame.

While I'm here, refactor the script to potentially support the fancy font in other languages someday. That's really just using a match statement instead of an if, though.... (and commenting it)