Use prefab add vert
This one is weirder...when I was testing, I found that recycling the same node over and over would cause confusing behavior when adding verts. The problem didn't occur when we were deleting and reconstructing the add button every time, so...just do that now, despite the premade node apparently being perfectly good as-is?