Comment and cleanup add-vert placement
For starters, identify what this section of code even is. Then the usual comment, clarify names, de-square the constant's value for ease of editing, etc. Trying hard to match existing logic behavior.

Unintuitively, "if mouse_position" is always true. If mouse position is ever null (is that possible?), the code which calculates nearest_position will crash before the code even reaches the if statement. In other words, nearest_position implies mouse_position, so it's valid logically to only check the former.