Formal API for de/selecting things
I can't imagine that deleting terrain within the polygon editor will be the only time we need to change selections from the outside in the future. Even if it isn't, having all the actual select-deselect code in the same place can't possibly hurt when it comes to understanding the system.

(also, add used parameters to the signature of the selection_changed signal. oh no signal signature)