Add info field to saved levels
Most of this is hardly critical yet, but. As I was updating the string IDs, I noticed I was breaking compatibility with the levels @InputZ already saved. I've always wanted to preserve compatibility across different game versions as much as possible, even through major format rewrites, and that starts here: saving the level format version as part of the level itself.