Update changelog, de-bump version number
With the position change gone, the only actual breaking change is introducing SemVer itself. Everything else is compatible between 0 and 0.0.1. It hardly seems appropriate to play it up as a bump--so I won't. It's version 0.0.0 now.

Notably, my v0.0.1 test file opens fine still. I should note that this is not a mistake in my eyes; I want forwards compatibility whenever possible, so as long as there's no new features in use, files made in newer versions *should* open in older versions, and without fanfare.