Load entire level as a parallel nodetree
The individual section-load functions were doing this separately, on a small scale, and it seemed like a good idea...and so I'm expanding it to the level-wide scale! The entire level is now loaded as a parallel tree, so that if anything goes wrong, it's easy to revert to the previous level and just say "level loading failed, haha!"

In the process, redesign loading to take Main instead of the level template--since with the nodetree being created in parallel, getting Main is all the level template is really used for.