Deselect all on level loaded
Aaagh my brain is not working to explain things :confounded:

Selection. List of nodes. Loading a level queue_free()s all nodes from the previous level, so if there's a selection when level is loaded, the selection becomes a list of dangling pointers, and clicking anywhere (even on an item? can't remember) immediately crashes due to use after free. Fixed by doing a deselect-all before the level is freed.

This bugfix is why I added all the signals in serializer. Genuinely the whole reason. Hopefully they're useful for other things.