Fix saving after loading by instantiating template
Was hitting a bug where saving a level after loading a level would crash due to a missing node. The load procedure wasn't creating the full Template node tree--it was skipping parents for unused polygon types. I knew that would be a problem eventually, but not this soon!

Refactor new-node-tree creation to instantiate a full blank template instead of creating from scratch. This fixes the bug, but by adhering to a single source of truth, should also reduce the future work needed to ensure save. load, and other editor code stay using compatible node trees.