Move placed verts via drag, not click
Keeping this commit light, and there's STILL a bit to unpack! The gist of it is, placed verts now respond to left clicks on press, so it feels like dragging and dropping. Changed their mouse icon to make that clearer, too.

Implementing this required a new script for placed verts, so that right click (to delete) could still respond to release only. (Because each button is now specially handled, the button action mode is now fully ignored.)

Kind of awesome that I didn't have to modify polygon_editor or polygon_draw to end drag on mouse up--it just Already Did That.