Show the edited polygon in realtime
Interesting that the body pattern shifts on finalizing edits. Which I think is because the origin is changed when the polygon is totally rebuilt.... We're not even planning to save polygon origins, are we? I think I can fix that by just originating all polygons at (0,0)....(or just at snapped intervals?)

Includes an on/off switch just in case.