Don't ignore camera zoom in the background
This is the easiest way to fix the main bug, where sizing the window up would not scale backgrounds properly and the "backstage" area would show. It also changes how zooming the camera looks--zooming in now zooms the background, where before it would keep the background its normal size.... It positions it weirdly too, but that's fixable.