Matching boolean to standard convention
Pretty self explanatory, I don't know why booleans were handled differently but it was an easy change so I digress.

Instead of using an AnimatedSprite2D for the whole button I went with just a simple Sprite2D for the checkmark alone, and then have the actual button be the background panel, which to me just makes more sense. The checkmark *can* be switched to an animated sprite if there are plans to animate the checkmark but this approach has the same effect as before.