Create custom container type for property list
The problem with how the property menu currently does things is, the widgets are each only aligning against their own labels, ignoring how the other widgets are aligned. This makes the property menu look jumbled.

This container is meant to fix that by aligning all widgets into the same column (and consequently the labels into a column as well). The actual alignment code is not yet written, but I've got pretty much everything else I can.