First non-crashing impl of labeled list draw
Doesn't quite work right--doesn't resize to fit contents--but it shows the properties, and it shows the labels. Plus one; that's good, right? 🙃

Turns out cancel_free is not technically supposed to be used. Huh. There goes that idea. But removing the child works better--simplifies the code, no need for extra branches I think.