Update invalidity marking during poly creation
Having the first line in invalid red is quite alarming and quite unnecessary; and it is, in fact, possible to have >3 verts during vertex creation, meaning validation *does* need to be done.

Since it's the same validation code as during dragging, let's just update both to use the same block of code.