Set field names instead of labels
To hook into the labeled-list container's behavior. Labels are soon to be no longer part of the field widgets at all...

Prop menu now crashes on opening twice. Wait no it doesn't--what the heck?! (For reference, error was in indexing a dictionary with value "AAAAA," which is the default value of the fields' integrated labels.... the item I tested on was a shine sprite.)